requirements.txt
tensorflow
streamlit
numpy
Pillow
scikit-learn
